README: re-organize elements
The documentation link is too close to the screenshot and the eyes think the URL
is part of the screenshot making many people missing the install documentation
link.

It would be better to design the README to put the description and the
documentation altogether and to put the screenshot at the very end of the
document to keep thing organized and more intuitive.